Move identity retrival code to common to use from oneshots
