+++ common/js/src/index.ts
@@ -7,3 +7,4 @@ export * from "./utils"
 export * from "./interfaces"
 export * from "./contract"
 export * from "./env"
+export * from "./identity"
